+++ src/main/java/com/example/edu/StartApp.java
@@ -1,0 +1,10 @@
+package com.example.edu;
+
+import com.example.edu.presentation.Main;
+
+public class StartApp {
+
+    public static void main(String[] args) {
+        new Main().main(args);
+    }
+}
