+++ src/main/java/com/example/edu/presentation/Main.java
@@ -1,14 +1,15 @@
-package com.example.edu;
+package com.example.edu.presentation;
 
 import java.io.File;
 import java.io.FileInputStream;
 
+import com.example.edu.domain.Schiffe;
+import com.example.edu.presentation.Controller;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.stage.Stage;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.layout.BorderPane;
 import javafx.scene.text.Font;
 
 
@@ -20,8 +21,8 @@ public class Main extends Application {
 	public void start(Stage primaryStage) {
 		try {
 			Font myFont = null;
-			myFont = Font.loadFont(new FileInputStream(new File("src/main/resources/com/example/edu/Connection-arMJ.otf")), 10);
-			//Parent root = FXMLLoader.load(getClass().getResource("Leaderboard.fxml"));	
+			myFont = Font.loadFont(new FileInputStream(new File("src/main/resources/com/example/edu/presentation/Connection-arMJ.otf")), 10);
+			//Parent root = FXMLLoader.load(getClass().getResource("Leaderboard.fxml"));
 			//Parent root = FXMLLoader.load(getClass().getResource("Start.fxml"));
 			Parent root = FXMLLoader.load(getClass().getResource("Karte.fxml"));
 			Controller.declarationClassement ();
@@ -32,11 +33,11 @@ public class Main extends Application {
 			primaryStage.show();
 
 			Schiffe test = new Schiffe ();
-			
+
 			test.initializeMap();
-			
-			//Schiffe point = new Schiffe (0, 0, "bas", 1); 
-			
+
+			//Schiffe point = new Schiffe (0, 0, "bas", 1);
+
 			Schiffe porteAvion = new Schiffe(5, '5');
 			Schiffe croiseur = new Schiffe(4, '4');
 			Schiffe contreTorpilleur = new Schiffe (3, 'A');
@@ -44,7 +45,7 @@ public class Main extends Application {
 			Schiffe torpilleur = new Schiffe (2, '2');
 
 			test.printMap();
-			
+
 		} catch(Exception e) {
 			e.printStackTrace();
 		}
