+++ src/main/resources/com/example/edu/presentation/Karte.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="593.0" prefWidth="780.0" styleClass="bodybg" stylesheets="@farben.css" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.edu.Controller">
+<AnchorPane prefHeight="593.0" prefWidth="780.0" styleClass="bodybg" stylesheets="@farben.css" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.edu.presentation.Controller">
 
 
    <children>
@@ -180,7 +180,7 @@
       <Label fx:id="lblVie4" layoutX="651.0" layoutY="194.0" text="4" AnchorPane.rightAnchor="120.0" AnchorPane.topAnchor="194.0" />
       <Label fx:id="lblVieA" layoutX="648.0" layoutY="263.0" text="3" AnchorPane.rightAnchor="123.0" AnchorPane.topAnchor="263.0" />
       <Label fx:id="lblVie5" layoutX="648.0" layoutY="338.0" text="5" AnchorPane.rightAnchor="123.0" AnchorPane.topAnchor="338.0" />
-     
+
       <Label fx:id="t35" layoutX="202.0" layoutY="549.0" prefHeight="27.0" prefWidth="53.0" text="/ 35 " AnchorPane.bottomAnchor="17.0" AnchorPane.rightAnchor="525.0" />
       <ImageView fx:id="marin" fitHeight="27.0" fitWidth="120.0" layoutX="593.0" layoutY="51.0" pickOnBounds="true" preserveRatio="true" AnchorPane.rightAnchor="67.0" AnchorPane.topAnchor="51.0">
          <image>
