+++ src/main/resources/com/example/edu/presentation/Leaderboard.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane onMouseDragOver="#classement" prefHeight="500.0" prefWidth="669.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.edu.LeaderboardController">
+<AnchorPane onMouseDragOver="#classement" prefHeight="500.0" prefWidth="669.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.edu.presentation.LeaderboardController">
    <children>
       <Button fx:id="Restart" layoutX="149.0" layoutY="374.0" mnemonicParsing="false" onAction="#ToKarte" text="Restart" />
       <Button fx:id="Exit" layoutX="455.0" layoutY="374.0" mnemonicParsing="false" onAction="#Exit" text="Exit" />
