+++ src/main/resources/com/example/edu/presentation/Start.css
@@ -1,0 +1,26 @@
+.bodybg{
+	-fx-background-image: url("green2.jpeg");
+	-fx-background-size: 100% 100%;
+}
+
+#Start { 
+    -fx-background-color: transparent;
+    -fx-background-image: url("start1.png");
+    -fx-background-size: 90px;
+    -fx-background-repeat: no-repeat;
+    -fx-background-position: 100%
+}
+
+#Start:hover {
+    -fx-background-image: url("start2.png");
+    -fx-background-size: 90px;
+    -fx-background-repeat: no-repeat;
+    -fx-background-position: 100%
+}
+
+#Start:pressed {
+    -fx-background-image: url("start3.png"); 
+    -fx-background-size: 90px;
+    -fx-background-repeat: no-repeat;
+    -fx-background-position: 100%
+}
