+++ src/main/resources/com/example/edu/presentation/ban.png
