+++ src/main/resources/com/example/edu/presentation/green2.jpeg
