+++ src/main/resources/com/example/edu/presentation/porteavion.png
