+++ src/main/resources/com/example/edu/presentation/restart.png
