+++ src/main/resources/com/example/edu/presentation/restart2.png
