+++ src/main/resources/com/example/edu/presentation/start2.png
