+++ src/main/resources/com/example/edu/presentation/torpilleur.png
