+++ src/main/resources/com/example/edu/presentation/touch5.gif
